design: 검색결과 정렬 필터링 UI 구현 #6
